Fix for GetRenderingParameters
GetRenderingParameters(NameValueCollection collection) was recently
changed to perform a null-check on the SitecoreTypeConfiguration, and
load it if it was null. However, the remaining method that took the
parameters as a string bypassed this. This commit has changed this to go
via the updated method. This also is consistent with the behaviour of
the other 2 GetRenderingParameters methods.