Removes second extension method that had the same method signature as RenderImage(T model, . . .) but with a different number of optional parameters.
C# does not allow two methods to differ only by optional parameters, as it gets confused about which method you want to call when you do not specify any optional parameters in the call.

Additionally adds the outputHeightWidth optional parameter to the standard RenderImage that points at the page Model to bring it in-line with how the model injeced method works.